requirements.txt
opencv-python
numpy
pandas
torch
torchvision
streamlit
Pillow
matplotlib
scikit-learn
